lesson_7/requirements.txt
amqp==5.1.1
aniso8601==7.0.0
asgiref==3.5.2
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.0.4
django-ckeditor==6.4.0
django-debug-toolbar==3.4.0
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.13.1
Faker==13.15.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
importlib-metadata==4.12.0
iniconfig==1.1.1
kombu==5.2.4
Markdown==3.3.7
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
promise==2.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==6.1.2
pytz==2022.1
redis==4.3.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
